Deprecate sprintf
