+++ builtins/core_test.go
@@ -15,7 +15,7 @@ import (
 var sourceFile map[string]string
 
 // TestCoreDocs tests documentation has been written for core builtins
-func __TestCoreDocs(t *testing.T) {
+func TestCoreDocs(t *testing.T) {
 	count.Tests(t, 1)
 	test.Exists(t, gopath.Source([]string{"builtins"})+"docgen_test.go")
