+++ config/app.go
@@ -5,4 +5,4 @@ const AppName = "murex"
 
 // Version number of $SHELL
 // Format of version string should be "(major).(minor).(revision) DESCRIPTION"
-const Version = "1.7.1000 BETA"
+const Version = "1.8.0000 BETA"
