+++ gen/user-guide/reserved_vars_doc.yaml
@@ -93,7 +93,7 @@
     the full path to the running shell. The reason for defining a reserved variable
     is so that the shell path cannot be overridden.
 
-    ### MUREX_ARGS
+    ### `MUREX_ARGS` (json)
 
     This is TODO: [https://github.com/lmorg/murex/issues/304](Github issue 304)
   Synonyms:
